requirements.txt
# Module Dependencies
pillow; python_version < '3.6'
matplotlib; python_version < '3.6'
numpy; python_version < '3.6'
# Setup Dependency
rstparse==1.0; python_version < '3.6'